+++ src/Exceptions/InvalidArgumentsException.php
@@ -3,6 +3,6 @@ namespace Codeat3\FoaasClient\Exceptions;
 
 use Exception;
 
-class InvalidArgumentsException extends Exception
+final class InvalidArguments extends Exception
 {
 }