+++ src/Exceptions/InvalidResponseException.php
@@ -3,6 +3,6 @@ namespace Codeat3\FoaasClient\Exceptions;
 
 use Exception;
 
-final class InvalidMethodCall extends Exception
+final class InvalidResponse extends Exception
 {
 }