+++ src/FoaasClient.php
@@ -3,53 +3,83 @@
 namespace Codeat3\FoaasClient;
 
 use GuzzleHttp\Client;
-use Codeat3\FoaasClient\Exceptions\InvalidArgumentsException;
-use Codeat3\FoaasClient\Exceptions\InvalidMethodCallException;
+use Codeat3\FoaasClient\Response\XmlResponse;
+use Codeat3\FoaasClient\Response\HtmlResponse;
+use Codeat3\FoaasClient\Response\JsonResponse;
+use Codeat3\FoaasClient\Response\TextResponse;
+use Codeat3\FoaasClient\Exceptions\InvalidArguments;
+use Codeat3\FoaasClient\Exceptions\InvalidMethodCall;
 
 class FoaasClient
 {
-    const API_ENDPOINT = 'https://foaas.com';
     protected static $endpoints;
     protected static $guzzleClient;
 
-    protected $method;
-    protected $arguments;
+    protected $responseType;
+    protected static $responseAs;
 
-    protected $typeHeaderMap = [
-        'json' => 'application/json',
-        'nocache' => 'application/json',
-        'text' => 'text/plain',
-        'html' => 'text/html',
-        'xml' => 'application/xml',
+    protected $url;
 
+    protected $responseTypeMap = [
+        'xml' => XmlResponse::class,
+        'json' => JsonResponse::class,
+        'text' => TextResponse::class,
+        'html' => HtmlResponse::class,
     ];
 
-    public function guzzleClient($responseType = 'json')
+    protected $decencyLevelMap = [
+        'nil' => [
+            'search' => '',
+            'replace' => '',
+        ],
+        'low' => [
+            'search' => 'fuck',
+            'replace'=> 'f*ck',
+        ],
+        'medium' => [
+            'search' => 'fuck',
+            'replace'=> 'f*#k',
+        ],
+        'high' => [
+            'search' => 'fuck',
+            'replace'=> 'f*#$',
+        ],
+        'extreme' => [
+            'search' => 'fuck',
+            'replace'=> '!*#$',
+        ]
+    ];
+
+    protected $method;
+    protected $arguments;
+    protected $decencyLevel;
+
+    public function __construct(array $config = [])
     {
-        if ($responseType == 'nocache') {
-            return new Client([
-                'headers' => [
-                    'Accept' => $this->typeHeaderMap[$responseType]
-                ],
-            ]);
-        }
-        if (is_null(self::$guzzleClient)) {
-            self::$guzzleClient = new Client([
-                'headers' => [
-                    'Accept' => $this->typeHeaderMap[$responseType]
-                ],
-            ]);
-        }
-        return self::$guzzleClient;
+        $this->decencyLevel = $config['decency'] ?? null;
+        $this->responseType = $config['responseAs'] ?? null;
+
+        $this->responseTypeMap = array_merge($this->responseTypeMap, ResponseFormatValidator::validate($config['responseFormats'] ?? null));
     }
 
-    private function buildUrl($method, array $fields = []) : string
+    private function getResponseType()
     {
-        $url = '/' . $method;
-        if (count($fields) > 0) {
-            $url .= '/' . implode('/', $fields);
+        if (is_null(self::$responseAs)) {
+            if (array_key_exists($this->responseType, $this->responseTypeMap)) {
+                self::$responseAs = new $this->responseTypeMap[$this->responseType]();
+            }
+            return new $this->responseTypeMap['json']();
         }
-        return self::API_ENDPOINT . $url;
+        return self::$responseAs;
+    }
+
+    public function guzzleClient()
+    {
+        return new Client([
+            'headers' => [
+                'Accept' => $this->getResponseType()->getHeaders()
+            ],
+        ]);
     }
 
     public function __call($method, $arguments)
@@ -57,13 +87,14 @@ class FoaasClient
         // check if the method is supported
         if (array_key_exists($method, $this->getAvailableEndpoints())) {
             $argumentsExpected = $this->getAvailableEndpoints()[$method]['fields'];
+            self::$guzzleClient = null;
             if (count($arguments) === count($argumentsExpected)) {
-                $this->url = $this->buildUrl($method, $arguments);
+                $this->url = UrlBuilder::buildUrl($method, $arguments);
             } else {
-                throw new InvalidArgumentsException();
+                throw new InvalidArguments();
             }
         } else {
-            throw new InvalidMethodCallException();
+            throw new InvalidMethodCall();
         }
         return $this;
     }
@@ -71,7 +102,7 @@ class FoaasClient
     public function getAvailableEndpoints() : array
     {
         if (is_null(self::$endpoints)) {
-            $response = json_decode($this->apiCall($this->buildUrl('operations'), 'nocache'), true);
+            $response = json_decode($this->apiCall(UrlBuilder::buildUrl('operations')), true);
             foreach ($response as $operation) {
                 $arrUrl = array_filter(explode("/", $operation['url']));
                 $endpoint = $arrUrl[1];
@@ -85,29 +116,42 @@ class FoaasClient
         return self::$endpoints;
     }
 
-    public function apiCall($path, $responseType = 'json')
+    public function apiCall($path)
     {
-        $response = $this->guzzleClient($responseType)->get($path);
+        $response = $this->guzzleClient()->get($path);
         return $response->getBody()->getContents();
     }
 
-    public function asArray()
+    /**
+     * A method to get the response
+     *
+     * @return string
+     */
+    public function get():string
     {
-        return json_decode($this->apiCall($this->url, 'json'), true);
+        $apiResponse = $this->apiCall($this->url);
+        return $this->getResponseType()->response($apiResponse);
     }
 
-    public function asJson()
+    /**
+     * An alias to get the response as json string
+     *
+     * @return string
+     */
+    public function getAsJson():string
     {
-        return $this->apiCall($this->url, 'json');
+        $this->responseType = 'json';
+        return $this->get();
     }
 
-    public function asText()
+    /**
+     * An alias to get the response as XML string
+     *
+     * @return string
+     */
+    public function getAsXml():string
     {
-        return $this->apiCall($this->url, 'text');
-    }
-
-    public function asHtml()
-    {
-        return $this->apiCall($this->url, 'html');
+        $this->responseType = 'xml';
+        return $this->get();
     }
 }