+++ tests/FoaasTest.php
@@ -2,22 +2,12 @@
 
 namespace Codeat3\FoaasClient\Tests;
 
-use PHPUnit\Framework\TestCase;
-use Codeat3\FoaasClient\FoaasClient;
-
-class FoaasTest extends TestCase
+class FoaasTest extends FoaasTestCase
 {
-    protected $client;
-
-    public function setUp()
-    {
-        parent::setUp();
-        $this->client  = new FoaasClient();
-    }
 
     /** @test */
     public function is_api_working()
     {
-        $this->assertArrayHasKey('message', $this->client->version());
+        $this->assertArrayHasKey('message', $this->client->version()->asArray());
     }
 }