+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to `foaas-client` will be documented in this file
 
-## 1.0.0 - 201X-XX-XX
+## 1.0.0 - 2019-07-12
 
-- initial release
+##### Initial Release
+* a very basic wrapper around FOAAS api
+* support for response asText(), asJson(), asArray(), asXml(), asArray()
+* added exceptions on invalid method calls & invalid number of arguments passed